RF_04.000.05 Delete class
