--- src/test/java/school/redrover/Folder14Test.java
@@ -1,37 +1,0 @@
-package school.redrover;
-
-import org.openqa.selenium.By;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import school.redrover.runner.BaseTest;
-
-public class Folder14Test extends BaseTest {
-    private final static String FOLDER_NAME = "Test";
-
-    private void createFolder(String folderName) {
-        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
-        getDriver().findElement(By.cssSelector("#name")).sendKeys(folderName);
-        getDriver().findElement(By.className("com_cloudbees_hudson_plugins_folder_Folder")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-    }
-    @Test
-    public void testAddDescription () {
-        final String description = "Test123";
-
-        createFolder("Test");
-
-        getDriver().findElement(By.name("_.description")).sendKeys(description);
-        getDriver().findElement(By.name("Submit")).click();
-
-        Assert.assertEquals(getDriver().findElement(By.id("view-message")).getText(),description);
-    }
-    @Test
-    public void testClickPreview() {
-        createFolder(FOLDER_NAME);
-
-        getDriver().findElement(By.name("_.description")).sendKeys("description123");
-        getDriver().findElement(By.className("textarea-show-preview")).click();
-
-        Assert.assertEquals(getDriver().findElement(By.className("textarea-preview")).getText(),"description123");
-    }
-}
